backend/requirements.txt
# Backend dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.5.3
email-validator==2.1.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
stripe==7.10.0
python-dotenv==1.0.0
Pillow==10.2.0
httpx==0.26.0
